AuthQ/requirements.txt
fastapi
uvicorn[standard]
python-keycloak-client
python-jose[cryptography]
pydantic

# Shared Libraries
pyyaml
-e ./shared/q_auth_parser
